Save some memory by using shorts

+++ TaskRipper.Core/IterationResult.cs
@@ -5,14 +5,14 @@ namespace TaskRipper.Core
     [DebuggerDisplay("[Thread # {ThreadNumber} Iteration {Iteration}]: {Result}")]
     public class IterationResult<TResult>
     {
-        public IterationResult(int threadNumber, int iteration, TResult result)
+        public IterationResult(short threadNumber, int iteration, TResult result)
         {
             ThreadNumber = threadNumber;
             Iteration = iteration;
             Result = result;
         }
 
-        public int ThreadNumber { get; }
+        public short ThreadNumber { get; }
         public int Iteration { get; }
         public TResult Result { get; }
     }
